Added an action filter that validates a model state. Added Food entity data annotations

+++ CookBookAPI.Data/FoodDto.cs
@@ -6,13 +6,13 @@
 
         public string Description { get; set; }
 
-        public int Carbonates { get; set; }
+        public double Carbonates { get; set; }
 
-        public int Fats { get; set; }
+        public double Fats { get; set; }
 
-        public int Proteins { get; set; }
+        public double Proteins { get; set; }
 
-        public int Calories { get; set; }
+        public double Calories { get; set; }
 
         public bool IsVegetarian { get; set; }
     }
